pyproject.toml
[project]
name = "recruitment-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.5",
    "ipykernel>=7.1.0",
    "langchain>=0.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.2.12",
    "langchain-core>=1.2.6",
    "langchain-openai>=0.1.22",
    "langgraph>=0.2.35",
    "pydantic>=2.8.2",
    "pypdf>=4.2.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.52.2",
    "tiktoken>=0.7.0",
]
